backend_main/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
eunjeon==0.4.0
idna==2.10
joblib==1.0.1
mysqlclient==2.0.3
numpy==1.20.1
pandas==1.2.3
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
urllib3==1.26.3
